Pie Charts Subclass

gpie class is a subclass for pie charts.

Details

This class is a subclass which show dialog boxes of pie charts for graphics editing.

Fields

top:

tkwin class object; parent of widget window.

alternateFrame:

tkwin class object; a special frame for some GUI parts.

vbbox1:

variableboxes class object; the frame to select variables.

vbbox2:

variableboxes class object; the frame to select facet variables.

lbbox1:

textfields class object; the frame to set axis labels and the main title.

tbbox1:

toolbox class object; the frame to set the font, the colour set, other option, and the theme.

Contains

NULL

Methods

plotWindow():

Create the window that make plots.

savePlot(plot):

Save the plot.

registRmlist(object):

Register deletable temporary objects.

removeRmlist():

Remove registered temporary objects.

setFront():

Set front parts of frames.

setBack():

Set back parts of frames.

getWindowTitle():

Get the title of the window.

getHelp():

Get the title of the help document.

getParms():

Get graphics settings parameters.

checkTheme(index):

Check themes.

checkVariable(var):

Check a variable length.

checkError(parms):

Check errors.

setDataframe(parms):

Set data frames.

getGgplot(parms):

Get ggplot.

getGeom(parms):

Get geom.

getScale(parms):

Get scale.

getCoord(parms):

Get coord.

getFacet(parms):

Get facet.

getXlab(parms):

Get xlab.

getYlab(parms):

Get ylab.

getZlab(parms):

Get zlab.

getMain(parms):

Get the main label.

getTheme(parms):

Get theme.

getOpts(parms):

Get other opts.

getPlot(parms):

Get the plot object.

getMessage():

Get the plot error message.

commandDoIt(command):

An wrapper function for command execution.

See also

Other plot: checkError,plot_base-method, checkTheme,plot_base-method, checkVariable,plot_base-method, commandDoIt,plot_base-method, factorize-class, gDensity-class, gbox-class, gcont-class, gdiscbar-class, gdist-class, getCoord,plot_base-method, getFacet,plot_base-method, getGeom,plot_base-method, getGgplot,plot_base-method, getHelp,plot_base-method, getMain,plot_base-method, getMessage,plot_base-method, getOpts,plot_base-method, getParms,plot_base-method, getPlot,plot_base-method, getScale,plot_base-method, getTheme,plot_base-method, getWindowTitle,plot_base-method, getXlab,plot_base-method, getYlab,plot_base-method, getZlab,plot_base-method, ghist-class, gkm-class, gline-class, gqq-class, gscat-class, gscatmat-class, plotWindow,plot_base-method, plot_base-class, registRmlist,plot_base-method, removeRmlist,plot_base-method, savePlot,plot_base-method, setBack,plot_base-method, setDataframe,plot_base-method, setFront,plot_base-method