Read text from clipboard

Read text from clipboard

read_clipboard()

Value

A string.

Examples

read_clipboard()

#> [1] "---\noutput: github_document\n---\n\n<!-- README.md is generated from README.Rmd. Please edit that file -->\n \n```{r, echo = FALSE}\nknitr::opts_chunk$set(\n  collapse = FALSE, fig.height = 4, fig.width = 6,\n  comment = \"#>\",\n  fig.path = \"README-\"\n)\n```\n\n***\n\n<!-- badges: start -->\n[![lifecycle](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://www.tidyverse.org/lifecycle/#experimental)\n[![CRAN status](https://www.r-pkg.org/badges/version/RcmdrPlugin.biostat)](https://cran.r-project.org/package=RcmdrPlugin.biostat)\n[![GitHub version](https://img.shields.io/badge/GitHub-v`r packageVersion(\"RcmdrPlugin.biostat\")`-brightgreen.svg)](https://github.com/GegznaV/RcmdrPlugin.biostat)\n[![AppVeyor build status](https://ci.appveyor.com/api/projects/status/hm4h2rjb8ayr2df1/branch/master?svg=true)](https://ci.appveyor.com/project/GegznaV/rcmdrplugin-biostat/branch/master)\n[![Travis-CI Build Status](https://travis-ci.org/GegznaV/RcmdrPlugin.biostat.png?branch=master)](https://travis-ci.org/GegznaV/RcmdrPlugin.biostat)\n[![Updated-on](https://img.shields.io/badge/Updated%20on-`r gsub('-', '--', Sys.Date())`-yellowgreen.svg)](/commits/master)\n<!-- badges: end -->\n\n***\n\n```{r logo-local, fig.align='center', out.height='30%', out.width='30%',echo=FALSE} \n knitr::include_graphics(\"https://raw.githubusercontent.com/GegznaV/RcmdrPlugin.biostat/master/docs/logo.png\") \n``` \n \n```{r include=FALSE}\n# fig_dir <- \"https://github.com/GegznaV/RcmdrPlugin.biostat/tree/master/inst/etc/fig/\"\nfig_dir <- \"https://raw.githubusercontent.com/GegznaV/RcmdrPlugin.biostat/master/inst/etc/fig/\"\n\nknitr::opts_chunk$set(fig.align  = \"center\")\n```\n\n<!-- \nAn *R Commander* Plugin for Basic (Bio)Statistical Routines  \n-->\n\nR Commander Plug-in for Basic Data Management and Analysis Tasks {#intro}\n==============================================================================\n\n**RcmdrPlugin.biostat** is an [**Rcmdr**](https://CRAN.R-project.org/package=Rcmdr) plug-in for the most common data wrangling, visualization and analysis tasks using **tidyverse** family functions as well as functions from other packages.\n\n\nDocumentation and more information available at http://gegznav.github.io/RcmdrPlugin.biostat\n\n\n## Install the package {#install}\n\nTo install the development version of the package from “GitHub”:\n\n```{r Install package, eval=FALSE}\nif (!\"devtools\" %in% installed.packages())  install.packages(\"devtools\")\ndevtools::install_github(\"GegznaV/RcmdrPlugin.biostat\")\n```\n\n## Load the plug-in correctly {#load}\n\nTo load the plug-in correctly, you should follow the following steps.\nYou can choose a manual or an automatic loading.\n\n\n\n### Load manually\n\n1. Never load the plug-in directly by loading package **RcmdrPlugin.biostat**.\nInstead, load package **Rcmdr** first either by using R code:\n\n```{r eval=FALSE}\nlibrary(Rcmdr)\n```\n\nOr by selecting **Rcmdr** in RStudio „Packages“ menu:\n```{r, echo=FALSE, fig.cap=CAPTION}\nknitr::include_graphics(paste0(fig_dir, \"01--load-rcmdr-in-rs.png\")) \n\nCAPTION = \"\" # Figure caption/description.\n```\n\n2. Then load the plug-in through `Rcmdr` menu bar:\n\n```{r, echo=FALSE, fig.cap=CAPTION}\nknitr::include_graphics(paste0(fig_dir, \"02--load-plugin--biostat.png\")) \n\nCAPTION = \"\" # Figure caption/description.\n```\n\n\n3. If `RcmdrPlugin.biostat` is loaded, you can access its functionality through `BioStat'19` menu in the menu bar:\n\n\n```{r, echo=FALSE, fig.cap=CAPTION}\nknitr::include_graphics(paste0(fig_dir, \"03--biostat-plugin-loaded.png\")) \n\nCAPTION = \"\" # Figure caption/description.\n```\n\n\n\n### Load automatically\n\nTo load **RcmdrPlugin.biostat** automatically, run the code:\n\n```{r eval=FALSE}\noptions(Rcmdr = list(plugins = c(\"RcmdrPlugin.biostat\"), console.output = FALSE))\nlibrary(Rcmdr)\n```\n\n### Activate *BioStat* (green) mode\n\nA more convenient way to use the plug-in is to activate the *BioStat* mode (green icon):\n\n```{r, echo=FALSE, fig.cap=CAPTION}\nknitr::include_graphics(paste0(fig_dir, \"04--activate-biostat-mode.png\")) \n\nCAPTION = \"\" # Figure caption/description.\n```\n\nIn this mode, additional buttons, which enable access to the most common functions, are added below the menu bar.\nThe meaning of these buttons:\n\n```{r, echo=FALSE, fig.cap=CAPTION}\nknitr::include_graphics(paste0(fig_dir, \"05--buttons-in-biostat-mode.png\")) \n\nCAPTION = \"\" # Figure caption/description.\n```\n\nNote: if active dataset is not selected, some buttons are inactive.\n\n\n### Close *BioStat* mode\n\nTo close this green mode, simply restart the *R Commander*: \n\n* R Commander menu bar ? BioStat'19 ? Session ? Restart R Commander\n\n\n## Install previous versions of **RcmdrPlugin.biostat**\n\nIf you need previous versions of **RcmdrPlugin.biostat**, you can install them from `GitHub`:\n\n- BioStat'19 (2019):\n```{r Install BioStat-19, eval=FALSE}\ndevtools::install_github(\"GegznaV/RcmdrPlugin.biostat\", ref = \"biostat19\")\n```\n\n\n- BioStat'18 (2018):\n```{r Install BioStat-18, eval=FALSE}\ndevtools::install_github(\"GegznaV/RcmdrPlugin.biostat\", ref = \"biostat18\")\n```\n\n\n***\n\n# The other useful packages {#related-packages}\n\nOther useful **Rcmdr** plug-ins:\n\na. **RcmdrPlugin.EZR.as.menu** -- an *R Commander* plugin for the most common statistical analyses (functionality is the same as in **RcmdrPlugin.EZR**, except that **RcmdrPlugin.EZR.as.menu** creates a separate \"EZR\" menu in a menu bar and does not reorganize the original **Rcmdr** menus in the menu bar);\nb. **RcmdrPlugin.KMggplot2** -- an *R Commander* plugin for **ggplot2** graphics.\n\n\nTo install these packages, use the following code:\n\n\n```{r Install other packages, eval=FALSE}\n# RcmdrPlugin.EZR.as.menu\ndevtools::install_github(\"GegznaV/RcmdrPlugin.EZR\", ref = \"ezr_as_menu\")\n\n# RcmdrPlugin.KMggplot2\ninstall.packages(\"RcmdrPlugin.KMggplot2\")\n```\n\nAfter the packages are installed, you can load them the automatically:\n\n```{r eval=FALSE}\noptions(Rcmdr = list(plugins = c(\n    \"RcmdrPlugin.KMggplot2\",\n    \"RcmdrPlugin.EZR.as.menu\",\n    \"RcmdrPlugin.biostat\",\n    NULL\n    ), \n    console.output = FALSE))\n\nlibrary(Rcmdr)\n```\n\n# Important\n\nSome of the functions in the package are based on and modified from functions in packages *Rcmdr*<!--, *RcmdrPlugin.KMggplot2*-->, and *RcmdrPlugin.EZR*.\nAll of those packages are under GPL-2/GPL-3 license.\n\n***\n\n\n\n\n<!-- * * * -->\n\n<!--  <p align=\"right\"> </p>     -->\n"