Checkboxes widget.

Checkboxes widget with command functions and tips when mouse is over the box.

bs_checkboxes(
  parent = top,
  boxes,
  labels = NULL,
  title = NULL,
  values = NULL,
  default_value = "0",
  commands = list(),
  default_command = function() { },
  tips = list(),
  default_tip = "",
  border = FALSE,
  layout = c("vertical", "horizontal"),
  sticky_buttons = "w",
  sticky_title = "w"
)

Arguments

parent
Parent frame.

boxes
(vector of strings) Variable names for each checkbox.

labels
(vector of strings) Labels for each checkbox.

title
Title for the set of checkboxes.

values
A vector of values ("0" or "1"). Overrides the `default_value`.

default_value
(string) Default value ("0" or "1").

commands
A named list of commands (functions) for checkbox. The names must match the values of "boxes".

default_command
(function) A default command.

tips
A named list of strings to be used as tips for checkbox. The names must match the values of "boxes".

default_tip
(string) a default tip.

border
(logical) Flag if the frame should have a border.

layout
(string) One of "vertical" (default) and "horizontal".

sticky_buttons
(string) `sticky` option for buttons.

sticky_title
(string) `sticky` option for title (if no border is used).

Value

A named list with fields `frame` (frame with the checkboxes), `var` (tcl/tk variables for each box), and `obj` (tcl/tk objects for each box).

Examples