Tcl/Tk text entry widget.

Tcl/Tk text entry widget.

bs_entry(
  parent = top,
  width = "28",
  value = "",
  label = "",
  label_position = c("left", "above", "right", "none"),
  label_color = getRcmdr("title.color"),
  padx = 0,
  pady = 0,
  sticky = "w",
  sticky_label = sticky,
  sticky_text = sticky,
  main_frame = tk2frame(parent),
  text_frame = tk2frame(main_frame),
  label_frame = tk2frame(main_frame),
  tip = "",
  label_tip = "",
  scroll_x = FALSE,
  on_click = do_nothing,
  on_double_click = do_nothing,
  on_triple_click = do_nothing,
  on_release = do_nothing,
  on_click_3 = do_nothing,
  on_double_click_3 = do_nothing,
  on_triple_click_3 = do_nothing,
  on_release_3 = do_nothing,
  on_key_release = do_nothing,
  use_context_menu = TRUE,
  bind_clear = TRUE,
  variable = tclVar(value),
  ...
)

# S3 method for bs_entry
get_values(obj, ..., trim = TRUE)

# S3 method for bs_entry
set_values(obj, values, ..., add = FALSE)

# S3 method for bs_entry
tk_get_state(obj, ...)

# S3 method for bs_entry
tk_disable(obj, ..., foreground = "grey")

# S3 method for bs_entry
tk_normalize(obj, ..., foreground = getRcmdr("title.color"))

# S3 method for bs_entry
tk_activate(obj, ..., foreground = getRcmdr("title.color"))

# S3 method for bs_entry
tk_read_only(obj, ..., foreground = getRcmdr("title.color"))

Arguments

parent

Parent Tcl/Tk window.

width

(numeric) Width of the widget.

value

(character) Initial value.

label

(character) Label of the widget.

label_position

(character) One of "left", "above", "right", "none"

label_color

(character)

padx

(numeric)

pady

(numeric)

sticky

(character)

sticky_label

(character)

sticky_text

(character)

main_frame

Tcl/Tk frame for all parts of the widget.

text_frame

Tcl/Tk frame for the entry widget.

label_frame

Tcl/Tk frame for the labels.

tip

(character) Text visible on hover.

label_tip

(character) Text visible on hover over the label.

scroll_x

(logical) Should scrollbar for x direction be added.

on_click

(function)

on_double_click

(function)

on_triple_click

(function)

on_release

(function)

on_click_3

(function)

on_double_click_3

(function)

on_triple_click_3

(function)

on_release_3

(function)

on_key_release

(function)

use_context_menu

(logical) Flag if the default context menu should be added

bind_clear

(logical) Bind "clear" on double right-click

variable

Tcl/Tk variable.

...

Other parameters sent to tcltk2::tk2entry()

trim

(logical) Should leading and tailing whitespace be trimmed in the result?

add

(logical) Should values be added to currently existing ones?

foreground

(string) Label text color.